+++ BackendBolsaDeTrabajoUTN/BackendBolsaDeTrabajoUTN/Models/AddStudentUniversityInfoRequest.cs
@@ -9,8 +9,8 @@
         public int SpecialtyPlan { get; set; }
         public int CurrentStudyYear { get; set; }
         public string StudyTurn { get; set; }
-        public int AverageMarksWithPostponement { get; set; }
-        public int AverageMarksWithoutPostponement { get; set; }
+        public decimal AverageMarksWithPostponement { get; set; }
+        public decimal AverageMarksWithoutPostponement { get; set; }
         public string CollegeDegree { get; set; }
     }
 }
